Separate slice files
Separate slice files.
