+++ src/LoginFormContainer.test.jsx
@@ -45,7 +45,7 @@ describe('LoginFormContainer', () => {
       });
 
       expect(dispatch).toBeCalledWith({
-        type: 'application/changeLoginField',
+        type: 'login/changeLoginField',
         payload: { name: 'email', value: 'new email' },
       });
     });
@@ -71,7 +71,7 @@ describe('LoginFormContainer', () => {
 
       fireEvent.click(getByText('Log out'));
 
-      expect(dispatch).toBeCalledWith({ type: 'application/logout' });
+      expect(dispatch).toBeCalledWith({ type: 'login/logout' });
     });
   });
 });
