+++ src/slice.js
@@ -1,147 +1,33 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { combineReducers } from 'redux';
 
-import { saveItem } from './services/storage';
+import restaurantReducer, {
+  setRegions,
+  setCategories,
+  selectRegion,
+  selectCategory,
+  setRestaurants,
+  setRestaurant,
+  loadInitialData,
+  loadRestaurants,
+  loadRestaurant,
+} from './slice/restaurantSlice';
 
-import {
-  fetchRegions,
-  fetchCategories,
-  fetchRestaurants,
-  fetchRestaurant,
-  postLogin,
-  postReview,
-} from './services/api';
+import reviewReducer, {
+  changeReviewField,
+  clearReviewFields,
+  setReviews,
+  loadReview,
+  sendReview,
+} from './slice/reviewSlice';
 
-import { equal } from './utils';
+import loginReducer, {
+  setAccessToken,
+  changeLoginField,
+  logout,
+  requestLogin,
+} from './slice/loginSlice';
 
-const initialReviewFields = {
-  score: '',
-  description: '',
-};
-
-const initialState = {
-  regions: [],
-  categories: [],
-  restaurants: [],
-  restaurant: null,
-  selectedRegion: null,
-  selectedCategory: null,
-  loginFields: {
-    email: '',
-    password: '',
-  },
-  accessToken: '',
-  reviewFields: {
-    ...initialReviewFields,
-  },
-};
-
-const { actions, reducer } = createSlice({
-  name: 'application',
-  initialState,
-  reducers: {
-    setRegions(state, { payload: regions }) {
-      return {
-        ...state,
-        regions,
-      };
-    },
-
-    setCategories(state, { payload: categories }) {
-      return {
-        ...state,
-        categories,
-      };
-    },
-
-    setRestaurants(state, { payload: restaurants }) {
-      return {
-        ...state,
-        restaurants,
-      };
-    },
-
-    setRestaurant(state, { payload: restaurant }) {
-      return {
-        ...state,
-        restaurant,
-      };
-    },
-
-    selectRegion(state, { payload: regionId }) {
-      const { regions } = state;
-      return {
-        ...state,
-        selectedRegion: regions.find(equal('id', regionId)),
-      };
-    },
-
-    selectCategory(state, { payload: categoryId }) {
-      const { categories } = state;
-      return {
-        ...state,
-        selectedCategory: categories.find(equal('id', categoryId)),
-      };
-    },
-
-    changeLoginField(state, { payload: { name, value } }) {
-      return {
-        ...state,
-        loginFields: {
-          ...state.loginFields,
-          [name]: value,
-        },
-      };
-    },
-
-    setAccessToken(state, { payload: accessToken }) {
-      return {
-        ...state,
-        accessToken,
-      };
-    },
-
-    logout(state) {
-      saveItem('accessToken', '');
-      return {
-        ...state,
-        accessToken: '',
-      };
-    },
-
-    changeReviewField(state, { payload: { name, value } }) {
-      return {
-        ...state,
-        reviewFields: {
-          ...state.reviewFields,
-          [name]: value,
-        },
-      };
-    },
-
-    clearReviewFields(state) {
-      return {
-        ...state,
-        reviewFields: {
-          ...initialReviewFields,
-        },
-      };
-    },
-
-    setReviews(state, { payload: reviews }) {
-      const { restaurant } = state;
-
-      return {
-        ...state,
-        restaurant: {
-          ...restaurant,
-          reviews,
-        },
-      };
-    },
-  },
-});
-
-export const {
+export {
   setRegions,
   setCategories,
   setRestaurants,
@@ -154,79 +40,16 @@ export const {
   changeReviewField,
   clearReviewFields,
   setReviews,
-} = actions;
+  loadInitialData,
+  loadRestaurants,
+  loadRestaurant,
+  requestLogin,
+  loadReview,
+  sendReview,
+};
 
-export function loadInitialData() {
-  return async (dispatch) => {
-    const regions = await fetchRegions();
-    dispatch(setRegions(regions));
-
-    const categories = await fetchCategories();
-    dispatch(setCategories(categories));
-  };
-}
-
-export function loadRestaurants() {
-  return async (dispatch, getState) => {
-    const {
-      selectedRegion: region,
-      selectedCategory: category,
-    } = getState();
-
-    if (!region || !category) {
-      return;
-    }
-
-    const restaurants = await fetchRestaurants({
-      regionName: region.name,
-      categoryId: category.id,
-    });
-    dispatch(setRestaurants(restaurants));
-  };
-}
-
-export function loadRestaurant({ restaurantId }) {
-  return async (dispatch) => {
-    dispatch(setRestaurant(null));
-
-    const restaurant = await fetchRestaurant({ restaurantId });
-
-    dispatch(setRestaurant(restaurant));
-  };
-}
-
-export function requestLogin() {
-  return async (dispatch, getState) => {
-    const { loginFields } = getState();
-    const { email, password } = loginFields;
-
-    const accessToken = await postLogin({ email, password });
-
-    saveItem('accessToken', accessToken);
-
-    dispatch(setAccessToken(accessToken));
-  };
-}
-
-export function loadReview({ restaurantId }) {
-  return async (dispatch) => {
-    const restaurant = await fetchRestaurant({ restaurantId });
-
-    dispatch(setReviews(restaurant.reviews));
-  };
-}
-
-export function sendReview({ restaurantId }) {
-  return async (dispatch, getState) => {
-    const { accessToken, reviewFields: { score, description } } = getState();
-
-    await postReview({
-      accessToken, restaurantId, score, description,
-    });
-
-    dispatch(loadReview({ restaurantId }));
-    dispatch(clearReviewFields());
-  };
-}
-
-export default reducer;
+export default combineReducers({
+  restaurant: restaurantReducer,
+  review: reviewReducer,
+  login: loginReducer,
+});
